Extract find bar to find_bar.rs (refs #68)
Moves render_find_bar() out of main.rs into a dedicated find_bar.rs
module. Draws the top-right scrollback search window, runs queries
against the focused pane's scrollback, and navigates matches with
Enter / Shift+Enter / arrows.

Mirrors wezterm-gui's per-concern module layout under termwindow/.

No behavior change — pure mechanical refactor.
